+++ crates/config/src/format.rs
@@ -1,4 +1,4 @@
-use crate::error::{ConfigError, ParserError};
+use crate::errors::{ConfigError, ParserError};
 use miette::{NamedSource, SourceOffset, SourceSpan};
 use serde::{de::DeserializeOwned, Serialize};
