+++ crates/config/src/lib.rs
@@ -1,7 +1,7 @@
 #![allow(clippy::result_large_err)]
 
 mod config;
-mod error;
+mod errors;
 mod format;
 mod layer;
 mod loader;
@@ -21,7 +21,7 @@ pub mod merge;
 pub mod validate;
 
 pub use config::*;
-pub use error::*;
+pub use errors::*;
 pub use format::*;
 pub use layer::*;
 pub use loader::*;
